group exceptoions together and test manager halt conditional exceptionality
