+++ pypiper/exceptions.py
@@ -4,13 +4,62 @@ __author__ = "Vince Reuter"
 __email__ = "vreuter@virginia.edu"
 
 
-__all__ = ["PipelineError", "UnsupportedFiletypeException"]
+__all__ = ["PipelineError", "PipelineHalt", "IllegalPipelineDefinitionError",
+           "IllegalPipelineExecutionError", "MissingCheckpointError",
+           "UnknownPipelineStageError", "UnsupportedFiletypeException"]
+
 
 
 
 class PipelineError(Exception):
     """ General pipeline error. """
     pass
+
+
+
+class IllegalPipelineDefinitionError(PipelineError):
+    pass
+
+
+
+class IllegalPipelineExecutionError(PipelineError):
+    """ Represent cases of illogical start/stop run() declarations. """
+    pass
+
+
+
+class MissingCheckpointError(Exception):
+    """ Represent case of expected but absent checkpoint file. """
+
+    def __init__(self, checkpoint, filepath):
+        msg = "{}: '{}'".format(checkpoint, filepath)
+        super(MissingCheckpointError, self).__init__(msg)
+
+
+
+class UnknownPipelineStageError(Exception):
+    """
+    Triggered by use of unknown/undefined name for a pipeline stage.
+
+    :param stage_name: Name of the stage triggering the exception.
+    :type stage_name: str
+    :param pipeline: Pipeline for which the stage is unknown/undefined.
+    :type pipeline: Pipeline
+    """
+
+
+    def __init__(self, stage_name, pipeline=None):
+        message = stage_name
+        if pipeline is not None:
+            try:
+                stages = pipeline.stages()
+            except AttributeError:
+                # Just don't contextualize the error with known stages.
+                pass
+            else:
+                message = "{}; defined stages: {}". \
+                    format(message, ", ".join(map(str, stages)))
+        super(UnknownPipelineStageError, self).__init__(message)
